feat: serve cover art that is embedded into file
Make the `getCoverArt` endpoint extract and serve the coverart image from the audio file for items that don't specify an `album_id` or refer to an album that does not specify an `artpath`.
Correspondingly, make the `getSong` endpoint return the song ID as `coverArt` ID when the song does not specify an `album_id`.

Extracts the cover art into a temporary file using `ffmpeg` and lets flask serve it.
Also, moves the `getCoverArt` endpoint implementation from `album.py` into a separate file since it is used to serve both album and single song cover art now.